requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
timm>=0.9.0  # For EfficientNetV2 pretrained models

# Data handling and processing
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
albumentations>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.12.0

# Configuration & Logging
pyyaml>=5.4.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.10.0

# Utilities
tqdm>=4.62.0
rich>=13.0.0
scipy>=1.11.0

# Development tools (optional)
jupyter>=1.0.0
ipython>=7.0.0
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
